+++ cmd/cli-docs/main.go
@@ -3,8 +3,9 @@ package main
 import (
 	"log"
 
-	"github.com/spf13/cobra/doc"
 	"github.com/tektoncd/results/pkg/cli/cmd"
+
+	"github.com/spf13/cobra/doc"
 )
 
 func main() {
