+++ pkg/cli/dev/cmd/pipelinerun/pipelinerun.go
@@ -1,25 +1,22 @@
-package result
+package pipelinerun
 
 import (
 	"github.com/spf13/cobra"
-	"github.com/tektoncd/results/pkg/cli/flags"
+	"github.com/tektoncd/results/pkg/cli/dev/flags"
 )
 
 // Command initializes a cobra command for `pipelinerun` sub commands
 func Command(params *flags.Params) *cobra.Command {
 	cmd := &cobra.Command{
-		Use:     "result",
-		Aliases: []string{"r", "results"},
-		Short:   "Query Results",
+		Use:     "pipelinerun",
+		Aliases: []string{"pr", "pipelineruns"},
+		Short:   "Query PipelineRuns",
 		Annotations: map[string]string{
 			"commandType": "main",
 		},
 	}
 
-	cmd.AddCommand(
-		ListCommand(params),
-		describeCommand(params),
-	)
+	cmd.AddCommand(listCommand(params))
 
 	return cmd
 }
