+++ pkg/cli/dev/cmd/records/get.go
@@ -22,11 +22,11 @@ import (
 	"text/tabwriter"
 	"text/template"
 
-	"github.com/jonboulle/clockwork"
-	"github.com/tektoncd/results/pkg/cli/format"
+	"github.com/tektoncd/results/pkg/cli/dev/flags"
+	"github.com/tektoncd/results/pkg/cli/dev/format"
 
+	"github.com/jonboulle/clockwork"
 	"github.com/spf13/cobra"
-	"github.com/tektoncd/results/pkg/cli/flags"
 	pb "github.com/tektoncd/results/proto/v1alpha2/results_go_proto"
 )
