+++ pkg/cli/dev/config/testdata/config.yaml
